requirements.txt
absl-py==2.4.0
blinker==1.9.0
cffi==2.0.0
click==8.3.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.29.0
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.33
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.1
pillow==12.2.0
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
Werkzeug==3.1.8
